feat(receipts): Updates checking mechanisms and adds auditor to complete checks
Checking a receipt now goes through received receipt, this is to ensure the state gets updated correctly and the checks are done in order. The user now provides a configured receipt auditor to perform the checks. Received receipt will use the auditor if the requested check is valid, and then store the result.

BREAKING CHANGE:

Signed-off-by: Bryan Cole <bryan.cole@semiotic.ai>