feat(aggregator): basic API version mgmt
By API, here we mean specifically the TAP JSON-RPC API version, not the
library version.

Maintains a const list of supported API versions, as well as a const
list of API versions for which to warn of imminent deprecation. We
expect the client to periodically check if its expected API version is
deprecated.

For all functions, an expected API version is to be supplied as argument.
Each function will have to check if the number is supported, and return
an error otherwise.

Signed-off-by: Alexis Asseman <alexis@semiotic.ai>