feat: TAP Showcase: Added integration tests for `tap_manager` and `tap_aggregator`.
A new JSON-RPC server is introduced which uses the `tap_manager` struct to handle receiving, verifying, and storing receipts.
When the server receives a parameterizable threshold of receipts, it makes a aggregate request to a parameterizable `tap_aggregator` server.
The tests use the new JSON-RPC server, called an indexer in the tests.
An indexer client sends a series of receipts to an indexer server.
The tests are intended to validate the correct operation of the `tap_manager` instantiated with the specified adapters and the `tap_aggregator`.

Signed-off-by: Severiano Sisneros <severiano@semiotic.ai>